+++ scripts/release-app/CHANGELOG.md
@@ -6,18 +6,20 @@ All notable changes to StegPass will be documented in this file.
 
 ### Added
 
-- `Get Password` to the gui to make it MUCH easier to retreive passwords from images
-- This CHANGELOG to more easily document changes between versions
-- StegPass icon is now present in the application
+- `Get Password` to the gui to make it MUCH easier to retreive passwords from images.
+- This CHANGELOG to more easily document changes between versions.
+- StegPass icon is now present in the application.
 
 ### Changed
 
 - Complete change in the GUIs handling of the title bar and menu bar (created from scratch)
 - Users cannot access parts of the app until an account is created.
+- The UI layout for creating passwords.
 
 ### Fixed
 
 - A formatting bug in the `Add Password` form that moves the button off the screen.
+- A bug where saving passwords fails when not saving to password folder.
 
 ## [0.0.1-alpha] - 6/12/2024